// workflow_fe/src/lib/api/gen/types/WorkflowApprovalStepConfigResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type WorkflowApprovalStepConfigResponseDtoStepTypeEnum =
  | "REVIEW"
  | "APPROVAL"
  | "FINANCE_CHECK"
  | "HR_CHECK"
  | "MANAGEMENT_APPROVAL"
  | "FINAL_VERIFICATION";

export type WorkflowApprovalStepConfigResponseDtoAssigneeTypeEnum =
  | "ROLE"
  | "USER"
  | "REQUESTER_MANAGER"
  | "DEPARTMENT_HEAD"
  | "CUSTOM_FIELD_USER";

export type WorkflowApprovalStepConfigResponseDtoEscalationAssigneeTypeEnum =
  | "ROLE"
  | "USER"
  | "REQUESTER_MANAGER"
  | "DEPARTMENT_HEAD"
  | "CUSTOM_FIELD_USER";

export type WorkflowApprovalStepConfigResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  workflowApprovalRuleId: string;
  /**
   * @type number
   */
  stepOrder: number;
  /**
   * @type string
   */
  stepName: string;
  /**
   * @type string
   */
  stepType: WorkflowApprovalStepConfigResponseDtoStepTypeEnum;
  /**
   * @type string
   */
  assigneeType: WorkflowApprovalStepConfigResponseDtoAssigneeTypeEnum;
  /**
   * @type string
   */
  assigneeRoleSlug: string | null;
  /**
   * @type string
   */
  assigneeUserId: string | null;
  /**
   * @type string
   */
  assigneeFieldPath: string | null;
  /**
   * @type boolean
   */
  isRequired: boolean;
  /**
   * @type boolean
   */
  requiresComment: boolean;
  /**
   * @type boolean
   */
  canReject: boolean;
  /**
   * @type boolean
   */
  canReassign: boolean;
  /**
   * @type number
   */
  slaHours: number | null;
  /**
   * @type string
   */
  escalationAssigneeType: WorkflowApprovalStepConfigResponseDtoEscalationAssigneeTypeEnum | null;
  /**
   * @type string
   */
  escalationAssigneeRoleSlug: string | null;
  /**
   * @type string
   */
  escalationAssigneeUserId: string | null;
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};